requirements.txt
pandas
streamlit
matplotlib
seaborn
urlextract
wordcloud
emoji
plotly.express
